refactor (config): split config.go into different files
