+++ task_sender/pkg/task_sender.go
@@ -36,7 +36,7 @@ func NewTask(provingSystemId common.ProvingSystemId, DAPayload serviceManager.Al
 
 type TaskSender struct {
 	avsWriter      *chainio.AvsWriter
-	eigenDAConfig  *config.EigenDADisperserConfig
+	eigenDAConfig  *config.EigenDAConfig
 	celestiaConfig *config.CelestiaConfig
 }
 
@@ -45,7 +45,7 @@ const RetryInterval = 1 * time.Second
 func NewTaskSender(config *config.TaskSenderConfig, avsWriter *chainio.AvsWriter) *TaskSender {
 	return &TaskSender{
 		avsWriter:      avsWriter,
-		eigenDAConfig:  config.EigenDADisperserConfig,
+		eigenDAConfig:  config.EigenDAConfig,
 		celestiaConfig: config.CelestiaConfig,
 	}
 }
